backend/python/requirements.txt
opencv-python
cognitive_face
python-dotenv